src/pyproject.toml
[tool.poetry]
name = "filmes-backend"
version = "0.1.0"
description = ""
authors = ["GustavoSwDaniel <gustavodanieldetoledo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.0"
asyncpg = "^0.26.0"
dependency-injector = "^4.40.0"
uvicorn = "^0.18.3"
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
httpx = "^0.23.0"
psycopg2-binary = "^2.9.3"
pytest = "^7.1.3"
factory-boy = "^3.2.1"
pytest-asyncio = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
